Big ol' route restructure
