+++ packages/atproto-browser/app/at/[identifier]/[collection]/(rkey)/layout.tsx
@@ -1,0 +1,44 @@
+import { resolveIdentity } from "@/lib/atproto-server";
+import type { Params } from "./[rkey]/page";
+import { getHandle, getPds } from "@atproto/identity";
+import Link from "next/link";
+import { Suspense } from "react";
+import { DidSummary } from "@/app/at/_lib/did-components";
+
+export default async function Layout({
+  children,
+  params,
+}: {
+  children: React.ReactNode;
+  params: Params;
+}) {
+  const identityResult = await resolveIdentity(params.identifier);
+  if (!identityResult.success) {
+    return <div>{identityResult.error}</div>;
+  }
+  const didDocument = identityResult.identity;
+  const handle = getHandle(didDocument);
+  if (!handle) {
+    return <div>No handle found for DID: {didDocument.id}</div>;
+  }
+  const pds = getPds(didDocument);
+  if (!pds) {
+    return <div>No PDS found for DID: {didDocument.id}</div>;
+  }
+
+  return (
+    <div>
+      <details>
+        <summary>
+          Author: {handle} (
+          <Link href={`/at/${didDocument.id}`}>{didDocument.id}</Link>)
+        </summary>
+        <Suspense>
+          <DidSummary did={didDocument.id} />
+        </Suspense>
+      </details>
+
+      {children}
+    </div>
+  );
+}
