+++ packages/atproto-browser/app/at/_lib/collection.tsx
@@ -1,6 +1,7 @@
 "use client";
 
 import { listRecords } from "@/lib/atproto";
+import { getAtUriPath } from "@/lib/util";
 import { AtUri } from "@atproto/syntax";
 import Link from "next/link";
 import { Suspense, useState } from "react";
@@ -34,7 +35,7 @@ export function CollectionItems({
         const uri = new AtUri(record.uri);
         return (
           <li key={record.uri}>
-            <Link href={`/at?u=${record.uri}`}>{uri.rkey}</Link>
+            <Link href={getAtUriPath(new AtUri(record.uri))}>{uri.rkey}</Link>
           </li>
         );
       })}
