+++ packages/atproto-browser/app/at/_lib/uri-bar.tsx
@@ -1,19 +1,30 @@
 "use client";
 
 import Link from "next/link";
-import { useSearchParams } from "next/navigation";
+import { useParams, usePathname, useSearchParams } from "next/navigation";
 import { AtUriForm } from "../../aturi-form";
 
 export function UriBar() {
   const searchParams = useSearchParams();
+  const pathname = usePathname();
+  const params = useParams() as {
+    identifier?: string;
+    collection?: string;
+    rkey?: string;
+  };
+
+  const uri =
+    pathname === "/at"
+      ? searchParams.get("u") ?? undefined
+      : `at://${[params.identifier, params.collection, params.rkey]
+          .map((c) => c && decodeURIComponent(c))
+          .filter(Boolean)
+          .join("/")}`;
 
   return (
     <div style={{ display: "flex", gap: 12 }}>
       <Link href="/">🏠</Link>
-      <AtUriForm
-        defaultUri={searchParams.get("u") ?? undefined}
-        style={{ flexGrow: 1 }}
-      />
+      <AtUriForm defaultUri={uri} style={{ flexGrow: 1 }} />
     </div>
   );
 }
